+++ src/Audacia.Commands/Validation/ValidationProperty.cs
@@ -19,14 +19,18 @@ namespace Audacia.Commands.Validation
         /// <param name="propertyName">The name of the property being validated.</param>
         /// <param name="displayName">The user friendly name of the property.</param>
         /// <param name="value">The property value.</param>
-        public ValidationProperty(IValidationModel validationModel, string propertyName, string? displayName, TProperty value)
+        public ValidationProperty(
+            IValidationModel validationModel,
+            string propertyName,
+            string? displayName,
+            TProperty value)
         {
             _validationModel = validationModel;
             PropertyName = propertyName;
             DisplayName = displayName ?? propertyName.SplitCamelCase();
             Value = value;
         }
-        
+
         /// <summary>
         /// Adds a validation error to the property.
         /// </summary>
@@ -34,7 +38,7 @@ namespace Audacia.Commands.Validation
         public void AddError(string message) => _validationModel.AddModelError(PropertyName, message);
 
         /// <summary>
-        /// Gets a user friendly model name.
+        /// Gets a user-friendly model name.
         /// </summary>
         public string ModelName => _validationModel.ModelName;
 
@@ -44,7 +48,7 @@ namespace Audacia.Commands.Validation
         public string PropertyName { get; }
 
         /// <summary>
-        /// Gets a user friendly property name.
+        /// Gets a user-friendly property name.
         /// </summary>
         public string DisplayName { get; }
 
@@ -52,17 +56,19 @@ namespace Audacia.Commands.Validation
         /// Gets the property value.
         /// </summary>
         public TProperty Value { get; }
-        
+
         /// <summary>
         /// Creates a child validation property on a property object.
         /// </summary>
         /// <typeparam name="TChild">Type of the member being validated.</typeparam>
-        /// <param name="property">Expresssion referencing the member being validated.</param>
-        /// <param name="displayName">User friendly member name.</param>
+        /// <param name="property">Expression referencing the member being validated.</param>
+        /// <param name="displayName">User-friendly member name.</param>
         /// <returns>A validation property.</returns>
         /// <exception cref="ArgumentNullException"><paramref name="property"/> is <see langword="null"/>.</exception>
-        public ValidationProperty<TChild> Property<TChild>(Expression<Func<TProperty, TChild>> property, string? displayName = null)
-        { 
+        public ValidationProperty<TChild> Property<TChild>(
+            Expression<Func<TProperty, TChild>> property,
+            string? displayName = null)
+        {
             if (property == null) throw new ArgumentNullException(nameof(property));
 
             var propertyName = GetNameForProperty(property);