+++ CHANGELOG.md
@@ -1,32 +1,50 @@
 # Changelog
 
+## 2.0.0 - 2025-06-05
+
+- Modified Validation model `ToCommandResultAsync` to `ToCommandResult` as no need to be async operation.
+- Corrected typos in documentation.
+
 ## 1.1.3 - 2024-11-20
+
 ### Added
+
 - No new functionality added
 
 ### Changed
+
 - No functionality changed
 
 ### Fixed
+
 - Upgraded vulnerable versions ([40e414f](https://github.com/audaciaconsulting/Audacia.Commands/commit/40e414f7cb92706e7185f328bde93771df077c3f))
 
 ## 1.1.2 - 2024-09-05
+
 ### Added
+
 - CHANGELOG.md file added
 
 ### Changed
+
 - README.md updated to include CHANGELOG.md reference
 
 ## 1.1.1 - 2023-08-02
+
 ### Added
+
 - No new functionality added
 
 ### Changed
+
 - Preparation for making library open source
 
 ## 1.0.0 - 2020-04-30
+
 ### Added
+
 - No new functionality added
 
 ### Changed
+
 - Renamed build pipeline to utilise revision